# docs/swagger.yaml
basePath: /
...
info:
  contact: {}
  description: IoC Cognition Fabric Node service — shared memory routing and memory
    operations proxy.
  title: CFN Service API
  version: "1.0"
paths:
...
  /api/workspaces/{workspaceId}/multi-agentic-systems/{masId}/semantic-negotiation/decide:
    post:
      consumes:
      - application/json
      description: Advances an existing semantic negotiation session with agent replies.
      parameters:
      - description: Workspace ID
        in: path
        name: workspaceId
        required: true
        type: string
      - description: Multi-Agentic System ID
        in: path
        name: masId
        required: true
        type: string
      - description: Semantic negotiation decide request
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Negotiation step executed successfully
          schema:
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Session not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Advance semantic negotiation session
      tags:
      - semantic-negotiation
  /api/workspaces/{workspaceId}/multi-agentic-systems/{masId}/semantic-negotiation/start:
    post:
      consumes:
      - application/json
      description: Initiates a new semantic negotiation session with multiple agents.
      parameters:
      - description: Workspace ID
        in: path
        name: workspaceId
        required: true
        type: string
      - description: Multi-Agentic System ID
        in: path
        name: masId
        required: true
        type: string
      - description: Semantic negotiation start request
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Negotiation session started successfully
          schema:
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Start semantic negotiation session
      tags:
      - semantic-negotiation
...
swagger: "2.0"